pyproject.toml
[project]
name = "rag-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.59",
    "langchain-google-genai>=2.1.4",
    "langchain-huggingface>=0.2.0",
    "langchain-pinecone>=0.2.6",
    "langchain-text-splitters>=0.3.8",
    "pdfplumber>=0.11.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
